+++ lib/src/main/java/com/orange/ods/compose/component/button/OdsButtonsCommon.kt
@@ -15,6 +15,7 @@ import androidx.compose.foundation.layout.size
 import androidx.compose.material.ButtonDefaults
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -53,6 +54,47 @@ class OdsButtonIcon : OdsComponentIcon {
     override fun Content(modifier: Modifier) {
         super.Content(modifier = modifier.size(ButtonDefaults.IconSize))
         Spacer(Modifier.size(ButtonDefaults.IconSpacing))
+    }
+}
+
+/**
+ * An icon in an [OdsIconButton].
+ */
+class OdsIconButtonIcon : OdsComponentIcon {
+
+    /**
+     * Creates an instance of [OdsIconButtonIcon].
+     *
+     * @param painter Painter of the icon.
+     * @param contentDescription The content description associated to this [OdsIconButtonIcon].
+     */
+    constructor(painter: Painter, contentDescription: String) : super(painter as Any, contentDescription)
+
+    /**
+     * Creates an instance of [OdsIconButtonIcon].
+     *
+     * @param imageVector Image vector of the icon.
+     * @param contentDescription The content description associated to this [OdsIconButtonIcon].
+     */
+    constructor(imageVector: ImageVector, contentDescription: String) : super(imageVector as Any, contentDescription)
+
+    /**
+     * Creates an instance of [OdsIconButtonIcon].
+     *
+     * @param bitmap Image bitmap of the icon.
+     * @param contentDescription The content description associated to this [OdsIconButtonIcon].
+     */
+    constructor(bitmap: ImageBitmap, contentDescription: String) : super(bitmap as Any, contentDescription)
+
+    override val tint: Color
+        @Composable
+        get() = iconButtonTintColor(displaySurface = displaySurface)
+
+    @Composable
+    internal fun Content(enabled: Boolean, displaySurface: OdsDisplaySurface = OdsDisplaySurface.Default) {
+        this.enabled = enabled
+        this.displaySurface = displaySurface
+        Content()
     }
 }